Add debugger support
This commit adds two method to support debugger in the extension side. One of the method will validate the breakpoints and send the
validation info to the extension. The other method will generate the debug info as json that is needed in the extension side to send it
to the synapse debugger server.