Fix connector flow for older car plugin versions
From MI extension version 2.0 onwards, the connector flow has been
changed. The connectors will be added to the pom file and resolved
while building instead of adding the zip to the project. The user
need to use the latest car plugin to support this. This commit fixes the
issue for older car plugin versions by adding the connector zip to the
project when loading the connector.