Add imported connector path for deletion support
Currently the mediator list is only sending the artifactId and version
of the conector to give support to delete it from the project. This will
remove the dependecy with the given artifactid and version from the pom.
But for imported connectors, there will be no pom entry. So, we need to
remove it from the connectors folder. This commit adds the path of the
imported connector to the mediator list to support deletion of imported
connectors.